+++ src/app/(mobile-ui)/withdraw/manteca/page.tsx
@@ -23,16 +23,15 @@ import AmountInput from '@/components/Global/AmountInput'
 import { formatUnits, parseUnits } from 'viem'
 import type { TransactionReceipt, Hash } from 'viem'
 import { PaymentInfoRow } from '@/components/Payment/PaymentInfoRow'
-import { useMantecaKycFlow } from '@/hooks/useMantecaKycFlow'
-import { MantecaGeoSpecificKycModal } from '@/components/Kyc/InitiateMantecaKYCModal'
 import { useAuth } from '@/context/authContext'
-import { useWebSocket } from '@/hooks/useWebSocket'
 import { useModalsContext } from '@/context/ModalsContext'
 import Select from '@/components/Global/Select'
 import { SoundPlayer } from '@/components/Global/SoundPlayer'
 import { useQueryClient } from '@tanstack/react-query'
 import { captureException } from '@sentry/nextjs'
 import useKycStatus from '@/hooks/useKycStatus'
+import { useMultiPhaseKycFlow } from '@/hooks/useMultiPhaseKycFlow'
+import { SumsubKycModals } from '@/components/Kyc/SumsubKycModals'
 import { usePendingTransactions } from '@/hooks/wallet/usePendingTransactions'
 import { PointsAction } from '@/services/services.types'
 import { usePointsConfetti } from '@/hooks/usePointsConfetti'
@@ -68,7 +67,6 @@ export default function MantecaWithdrawFlow() {
     const [selectedBank, setSelectedBank] = useState<MantecaBankCode | null>(null)
     const [accountType, setAccountType] = useState<MantecaAccountType | null>(null)
     const [errorMessage, setErrorMessage] = useState<string | null>(null)
-    const [isKycModalOpen, setIsKycModalOpen] = useState(false)
     const [isDestinationAddressValid, setIsDestinationAddressValid] = useState(false)
     const [isDestinationAddressChanging, setIsDestinationAddressChanging] = useState(false)
     // price lock state - holds the locked price from /withdraw/init
@@ -78,11 +76,16 @@ export default function MantecaWithdrawFlow() {
     const { sendMoney, balance } = useWallet()
     const { signTransferUserOp } = useSignUserOp()
     const { isLoading, loadingState, setLoadingState } = useContext(loadingStateContext)
-    const { user, fetchUser } = useAuth()
+    const { user } = useAuth()
     const { setIsSupportModalOpen } = useModalsContext()
     const queryClient = useQueryClient()
-    const { isUserBridgeKycApproved } = useKycStatus()
+    const { isUserMantecaKycApproved } = useKycStatus()
     const { hasPendingTransactions } = usePendingTransactions()
+
+    // inline sumsub kyc flow for manteca users who need LATAM verification
+    const sumsubFlow = useMultiPhaseKycFlow({
+        regionIntent: 'LATAM',
+    })
     // Get method and country from URL parameters
     const selectedMethodType = searchParams.get('method') // mercadopago, pix, bank-transfer, etc.
     const countryFromUrl = searchParams.get('country') // argentina, brazil, etc.
@@ -106,28 +109,12 @@ export default function MantecaWithdrawFlow() {
         isLoading: isCurrencyLoading,
     } = useCurrency(selectedCountry?.currency!)
 
-    // Initialize KYC flow hook
-    const { isMantecaKycRequired } = useMantecaKycFlow({ country: selectedCountry })
-
     // validates withdrawal against user's limits
     // currency comes from country config - hook normalizes it internally
     const limitsValidation = useLimitsValidation({
         flowType: 'offramp',
         amount: usdAmount,
         currency: selectedCountry?.currency,
-    })
-
-    // WebSocket listener for KYC status updates
-    useWebSocket({
-        username: user?.user.username ?? undefined,
-        autoConnect: !!user?.user.username,
-        onMantecaKycStatusUpdate: (newStatus) => {
-            if (newStatus === 'ACTIVE' || newStatus === 'WIDGET_FINISHED') {
-                fetchUser()
-                setIsKycModalOpen(false)
-                setStep('review') // Proceed to review after successful KYC
-            }
-        },
     })
 
     // Get country flag code
@@ -200,14 +187,8 @@ export default function MantecaWithdrawFlow() {
         }
         setErrorMessage(null)
 
-        // check if we still need to determine KYC status
-        if (isMantecaKycRequired === null) {
-            return
-        }
-
-        // check KYC status before proceeding to review
-        if (isMantecaKycRequired === true) {
-            setIsKycModalOpen(true)
+        if (!isUserMantecaKycApproved) {
+            await sumsubFlow.handleInitiateKyc()
             return
         }
 
@@ -250,8 +231,9 @@ export default function MantecaWithdrawFlow() {
         usdAmount,
         currencyCode,
         currencyAmount,
-        isMantecaKycRequired,
+        isUserMantecaKycApproved,
         isLockingPrice,
+        sumsubFlow.handleInitiateKyc,
     ])
 
     const handleWithdraw = async () => {
@@ -342,7 +324,6 @@ export default function MantecaWithdrawFlow() {
         setSelectedBank(null)
         setAccountType(null)
         setErrorMessage(null)
-        setIsKycModalOpen(false)
         setIsDestinationAddressValid(false)
         setIsDestinationAddressChanging(false)
         setBalanceErrorMessage(null)
@@ -468,6 +449,7 @@ export default function MantecaWithdrawFlow() {
     }
     return (
         <div className="flex min-h-[inherit] flex-col gap-8">
+            <SumsubKycModals flow={sumsubFlow} />
             <NavHeader
                 title="Withdraw"
                 onPrev={() => {
@@ -650,22 +632,6 @@ export default function MantecaWithdrawFlow() {
                         {errorMessage && <ErrorAlert description={errorMessage} />}
                     </div>
 
-                    {/* KYC Modal */}
-                    {isKycModalOpen && selectedCountry && (
-                        <MantecaGeoSpecificKycModal
-                            isUserBridgeKycApproved={isUserBridgeKycApproved}
-                            isMantecaModalOpen={isKycModalOpen}
-                            setIsMantecaModalOpen={setIsKycModalOpen}
-                            onClose={() => setIsKycModalOpen(false)}
-                            onManualClose={() => setIsKycModalOpen(false)}
-                            onKycSuccess={() => {
-                                setIsKycModalOpen(false)
-                                fetchUser()
-                                setStep('review')
-                            }}
-                            selectedCountry={selectedCountry}
-                        />
-                    )}
                 </div>
             )}
