+++ src/components/Home/HomeHistory.tsx
@@ -13,9 +13,8 @@ import { twMerge } from 'tailwind-merge'
 import Card from '../Global/Card'
 import { type CardPosition, getCardPosition } from '../Global/Card/card.utils'
 import EmptyState from '../Global/EmptyStates/EmptyState'
-import { KycStatusItem } from '../Kyc/KycStatusItem'
+import { KycStatusItem, isKycStatusItem, type KycHistoryEntry } from '../Kyc/KycStatusItem'
 import { BridgeTosReminder } from '../Kyc/BridgeTosReminder'
-import { isKycStatusItem, type KycHistoryEntry } from '@/hooks/useBridgeKycFlow'
 import { useBridgeTosStatus } from '@/hooks/useBridgeTosStatus'
 import { useWallet } from '@/hooks/wallet/useWallet'
 import { BadgeStatusItem } from '@/components/Badges/BadgeStatusItem'
