--- src/components/Kyc/InitiateBridgeKYCModal.tsx
@@ -1,96 +1,0 @@
-import ActionModal from '@/components/Global/ActionModal'
-import { useBridgeKycFlow } from '@/hooks/useBridgeKycFlow'
-import IframeWrapper from '@/components/Global/IframeWrapper'
-import { KycVerificationInProgressModal } from './KycVerificationInProgressModal'
-import { type IconName } from '@/components/Global/Icons/Icon'
-import { saveRedirectUrl } from '@/utils/general.utils'
-import useClaimLink from '../Claim/useClaimLink'
-import { useEffect } from 'react'
-
-interface BridgeKycModalFlowProps {
-    isOpen: boolean
-    onClose: () => void
-    onKycSuccess?: () => void
-    onManualClose?: () => void
-    flow?: 'add' | 'withdraw' | 'request_fulfillment'
-}
-
-export const InitiateBridgeKYCModal = ({
-    isOpen,
-    onClose,
-    onKycSuccess,
-    onManualClose,
-    flow,
-}: BridgeKycModalFlowProps) => {
-    const {
-        isLoading,
-        error,
-        iframeOptions,
-        isVerificationProgressModalOpen,
-        handleInitiateKyc,
-        handleIframeClose,
-        closeVerificationProgressModal,
-        resetError,
-    } = useBridgeKycFlow({ onKycSuccess, flow, onManualClose })
-    const { addParamStep } = useClaimLink()
-
-    // Reset error when modal opens to ensure clean state
-    useEffect(() => {
-        if (isOpen) {
-            resetError()
-        }
-    }, [isOpen, resetError])
-
-    const handleVerifyClick = async () => {
-        // Only add step param for claim flows (not add-money flow which has its own URL state)
-        if (flow !== 'add') {
-            addParamStep('bank')
-        }
-        const result = await handleInitiateKyc()
-        if (result?.success) {
-            saveRedirectUrl()
-            onClose()
-        }
-    }
-
-    return (
-        <>
-            <ActionModal
-                visible={isOpen}
-                onClose={onManualClose ? onManualClose : onClose}
-                title="Verify your identity first"
-                description="To continue, you need to complete identity verification. This usually takes just a few minutes."
-                icon={'badge' as IconName}
-                modalPanelClassName="max-w-full m-2"
-                ctaClassName="grid grid-cols-1 gap-3"
-                ctas={[
-                    {
-                        text: isLoading ? 'Loading...' : 'Verify now',
-                        onClick: handleVerifyClick,
-                        variant: 'purple',
-                        disabled: isLoading,
-                        shadowSize: '4',
-                        icon: 'check-circle',
-                        className: 'h-11',
-                    },
-                    ...(error
-                        ? [
-                              {
-                                  text: error,
-                                  onClick: onClose,
-                                  variant: 'transparent' as const,
-                                  className:
-                                      'underline text-xs md:text-sm !font-normal w-full !transform-none !pt-2 text-error-3 px-0',
-                              },
-                          ]
-                        : []),
-                ]}
-            />
-            <IframeWrapper {...iframeOptions} onClose={handleIframeClose} />
-            <KycVerificationInProgressModal
-                isOpen={isVerificationProgressModalOpen}
-                onClose={closeVerificationProgressModal}
-            />
-        </>
-    )
-}
