+++ src/hooks/useSumsubKycFlow.ts
@@ -80,6 +80,30 @@ export const useSumsubKycFlow = ({ onKycSuccess, onManualClose, regionIntent }: 
         fetchCurrentStatus()
     }, [regionIntent])
 
+    // polling fallback for missed websocket events.
+    // when the verification progress modal is open, poll status every 5s
+    // so the flow can transition even if the websocket event never arrives.
+    useEffect(() => {
+        if (!isVerificationProgressModalOpen) return
+
+        const pollStatus = async () => {
+            try {
+                const response = await initiateSumsubKyc({
+                    regionIntent: regionIntentRef.current,
+                    levelName: levelNameRef.current,
+                })
+                if (response.data?.status) {
+                    setLiveKycStatus(response.data.status)
+                }
+            } catch {
+                // silent — polling is a best-effort fallback
+            }
+        }
+
+        const interval = setInterval(pollStatus, 5000)
+        return () => clearInterval(interval)
+    }, [isVerificationProgressModalOpen])
+
     const handleInitiateKyc = useCallback(
         async (overrideIntent?: KYCRegionIntent, levelName?: string) => {
             setIsLoading(true)
